# app/Models/Property.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;

class Property  extends Model
{
    use HasFactory;
    protected $fillable=[
        "name",
        "address",
        "price",
        "type",
        "number_units"
    ];
}